fix some bug
