+++ dinky-web/src/pages/RegCenter/Alert/AlertInstance/components/AlertTypeChoose/InstanceForm/Sms/function.tsx
@@ -407,42 +407,3 @@ export const renderCtyunForm = () => {
   </>;
 };
 
-
-export const MANU_FACTURERS = [
-  {
-    label: l('rc.ai.mf.alibaba'),
-    value: 1,
-  },
-  {
-    label: l('rc.ai.mf.huawei'),
-    value: 2,
-  },
-  {
-    label: l('rc.ai.mf.yunpian'),
-    value: 3,
-  },
-  {
-    label: l('rc.ai.mf.tencent'),
-    value: 4,
-  },
-  {
-    label: l('rc.ai.mf.uni'),
-    value: 5,
-  },
-  {
-    label: l('rc.ai.mf.jdcloud'),
-    value: 6,
-  },
-  {
-    label: l('rc.ai.mf.cloopen'),
-    value: 7,
-  },
-  {
-    label: l('rc.ai.mf.emay'),
-    value: 8,
-  },
-  {
-    label: l('rc.ai.mf.ctyun'),
-    value: 9,
-  },
-];
