+++ dinky-web/src/pages/RegCenter/Alert/AlertInstance/constans.ts
@@ -49,3 +49,42 @@ export const ALERT_TYPE_LIST_OPTIONS: DefaultOptionType[] = [
   }
 ];
 
+
+export const MANU_FACTURERS = [
+  {
+    label: l('rc.ai.mf.alibaba'),
+    value: 1,
+  },
+  {
+    label: l('rc.ai.mf.huawei'),
+    value: 2,
+  },
+  {
+    label: l('rc.ai.mf.yunpian'),
+    value: 3,
+  },
+  {
+    label: l('rc.ai.mf.tencent'),
+    value: 4,
+  },
+  {
+    label: l('rc.ai.mf.uni'),
+    value: 5,
+  },
+  {
+    label: l('rc.ai.mf.jdcloud'),
+    value: 6,
+  },
+  {
+    label: l('rc.ai.mf.cloopen'),
+    value: 7,
+  },
+  {
+    label: l('rc.ai.mf.emay'),
+    value: 8,
+  },
+  {
+    label: l('rc.ai.mf.ctyun'),
+    value: 9,
+  },
+];
